test: split domain parser tests from problem parser tests
